# Conflicts:
#	docs/reference/scalar-type-audit.md
#	docs/reference/turn-event-coverage.md

+++ src/game/navy/TOcean.cpp
@@ -135,7 +135,9 @@ void TOcean::ReadFrom(TStream* stream) {
   TObject::ReadFrom(stream);
   EnsureSelectedTaskForceForOrderOwnerAndRefresh(0);
 
-  int i;
+  // A short counter: the bound test is a 16-bit `cmp word ptr [eax], bx` (0x5623a3), so
+  // an int loop variable would need a movsx of nationCount on every compare.
+  short i;
   for (i = 0; i < nationCount; ++i) {
     TZone* zone = &contextArray[i];
     if (g_pMapActionContextListHead == zone) {
